refactor(headlamp-plugin): rename karta-util to kartaUtil
Matches the plugin's camelCase file naming convention.

Refs #302
