+++ headlamp-plugin/src/lib/karta/kartaUtil.ts
